feat: Add SpecifyCourseCohortView for assigning course cohorts to user courses
This commit adds the SpecifyCourseCohortView class to the leerplan_api/Course/views.py file. The view is responsible for assigning a course cohort to a user course. It handles the PATCH request and updates the user course's cohort based on the provided course cohort ID. This addition enhances the functionality of managing course cohorts for user courses.