refactor: Remove isLoading state from RightPane component
This commit removes the isLoading state from the RightPane component in the leerplan_interface/leerplan/src/components/RightPane.jsx file. The isLoading state was previously used to display a loading message while the courses were being fetched, but it is no longer necessary as the loading logic has been refactored. This change simplifies the code and improves the overall performance of the component.