refactor: Update PYDANTIC_CLASSES_FILE path to use absolute path
This commit updates the PYDANTIC_CLASSES_FILE path in the leerplan_api/DataSynthesis/synthesise.py file to use an absolute path instead of a relative path. By using the  module, the code now correctly constructs the absolute path to the  file. This change ensures that the file is always found regardless of the current working directory, improving the reliability and portability of the code.