refactor: Fix loading message display in RightPane component
This commit fixes the display of the loading message in the RightPane component of the leerplan_interface/leerplan/src/components/RightPane.jsx file. Previously, the loading message was not being shown when the courses were being fetched. This change ensures that the loading message is correctly displayed, improving the user experience and providing feedback to the user while the data is being loaded.