feat: Add DetermineTimeChunksView for generating free time slots
The DetermineTimeChunksView is added to the Course views module to handle the generation of free time slots for a user. This view takes a list of courses as input and retrieves the lecture days for those courses. It then processes the data to create free time slots based on the user's metadata, including wake time and sleep time. This feature enhances the functionality of the application by providing users with a convenient way to determine their available time slots for studying or other activities.