feat: Refactor DetermineTimeChunksView to handle missing user metadata
The DetermineTimeChunksView in the Course views module has been refactored to handle the case when user metadata is missing. If the user metadata does not exist, a proper error response is returned. This refactor improves the robustness of the application and ensures that the user metadata is required for generating time blocks.