refactor: Fix conditional rendering of UserMetaData component
The code in UserDetailCard.jsx has been refactored to fix the conditional rendering of the UserMetaData component. Previously, the component was only rendered if the  object existed, but now it will also be rendered if the  property exists. This refactor ensures that the UserMetaData component is displayed correctly and improves the overall functionality of the UserDetailCard component.