refactor: Update AddUserMetaData component to validate input values
The AddUserMetaData component in AddUserMetaData.jsx has been updated to include input value validation. The getInputBorderColor function has been added to dynamically set the border color of the input fields based on the input values and regular expressions. This refactor improves the user experience by providing visual feedback on the validity of the input values.