refactor: Add profile picture upload path and adjust_time function
This commit adds the profile_picture_upload_path function to helper.py, which returns the path where the profile picture will be stored. It also includes the adjust_time function, which adjusts the time from \00:00:00\ to \23:59:59\ for easier time comparison. These changes improve the functionality and accuracy of the Account module.