refactor: Update api_urls.js to include GENERATE_TIME_BLOCKS_URL
The api_urls.js file in the leerplan_interface/leerplan/src/utility directory has been updated to include the GENERATE_TIME_BLOCKS_URL constant. This change allows for the generation of time blocks for a course through the specified URL. It enhances the functionality of the utility module by providing a convenient way to generate time blocks for courses.