refactor: Add EVENT_BREAK constant for event duration
The constants.js file in the leerplan_interface/leerplan/src/utility directory has been updated to include the EVENT_BREAK constant, which represents the duration of an event break. This change enhances the functionality of the utility module by providing a standardized value for event breaks, improving the accuracy of event scheduling.