feat: Add constant for displaying time blocks
The constants.js file in the leerplan_interface/leerplan/src/utility directory has been updated to include the DISPLAY_TIME_BLOCKS constant. This change allows for the display of time blocks on the calendar. It enhances the functionality of the utility module by providing a convenient way to show time blocks, improving the visual representation of the calendar.