feat: Update generateEvents function to include displayTimeBlocks parameter
The generateEvents function in CalendarEvents.jsx has been updated to include a new parameter, displayTimeBlocks. This change allows for the conditional generation of time block events based on the value of displayTimeBlocks. It enhances the flexibility of the function by providing control over whether time blocks should be displayed on the calendar.