feat: Add displayTimeBlocks prop to RightPane component
The RightPane component in RightPane.jsx has been updated to include a new prop, displayTimeBlocks. This change allows for the conditional display of time blocks based on the value of displayTimeBlocks. It enhances the functionality of the component by providing control over whether time blocks should be shown, improving the customization of the user interface.