refactor: Add generateICSFile function to create ICS files for events
This commit adds a new utility function, generateICSFile, to the ics_generator.js file. The function takes an array of events and generates an ICS file using the createEvents function from the ics package. It also includes logic to format the event details, such as the description, start and end times, location, and alarms. This addition enhances the functionality of the leerplan_interface by allowing users to generate ICS files for their events.